requirements.txt
flask
flask-sqlalchemy
flask-migrate
werkzeug
googlemaps
gmplot
ortools
numpy 
flask-wtf
flask-login
python-dotenv
mysqlclient==1.3.3
pymysql